refactor: refatorando o runrunit service. Agora ele está no shared pois será usado em outros dominios
