+++ src/shared/infrastructure/runrunit/runrunit.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { RunrunitService } from "./runrunit.service";
-import { EnvConfigModule } from "../../../../shared/infrastructure/env-config/env-config.module";
+import { EnvConfigModule } from "../env-config/env-config.module";
 
 @Module({
     imports: [EnvConfigModule],
