+++ src/tasks/application/usecases/end-work.usecase.ts
@@ -2,8 +2,8 @@ import { BadRequestError } from "../../../shared/application/errors/bad-request-
 import { UseCase as DefaultUseCase } from "../../../shared/application/usecases/use-case";
 import { UserRepository } from "../../../users/domain/repositories/user.repository";
 import { TaskEntity } from "../../domain/entities/task.entity";
-import { PauseTaskDto } from "../../infrastructure/external/runrunit/dtos/pause-task.dto";
-import { IRunrunitRepository } from "../../infrastructure/external/runrunit/repositories/i-runrunit-repository";
+import { PauseTaskDto } from "../../../shared/infrastructure/runrunit/dtos/pause-task.dto";
+import { IRunrunitRepository } from "../../../shared/infrastructure/runrunit/repositories/i-runrunit-repository";
 import { DefaultOutput, DefaultOutputMapper } from "../dtos/default-output.dto";
 
 export namespace EndWork {
