+++ src/tasks/application/usecases/estimate-task.usecase.ts
@@ -2,8 +2,8 @@ import { UseCase as DefaultUseCase } from "../../../shared/application/usecases/
 import { EstimateTaskAgentText } from "../../../shared/infrastructure/templates/estimate-task-agent-text";
 import { DefaultInputDto } from "../../infrastructure/external/ia/dtos/default-input.dto";
 import { IIARepository } from "../../infrastructure/external/ia/repositories/i-ia-repository";
-import { GetTaskDto } from "../../infrastructure/external/runrunit/dtos/get-task.dto";
-import { IRunrunitRepository } from "../../infrastructure/external/runrunit/repositories/i-runrunit-repository";
+import { GetTaskDto } from "../../../shared/infrastructure/runrunit/dtos/get-task.dto";
+import { IRunrunitRepository } from "../../../shared/infrastructure/runrunit/repositories/i-runrunit-repository";
 import { DefaultOutput, DefaultOutputMapper } from "../dtos/default-output.dto";
 
 export namespace EstimateTask {
